Setup hierarchical structure in codebase depending on functions
