+++ convert/file/docx2txt.go
@@ -1,4 +1,4 @@
-package convert
+package file
 
 import (
 	"os"
