+++ download/list/url_list_test.go
@@ -1,4 +1,4 @@
-package download
+package list
 
 import (
 	"io/ioutil"
