+++ review/logic/doc.go
@@ -1,4 +1,4 @@
 // Package review defines the core logic for managing and executing workflows within the application.
 // It handles the review processes, error handling, and the orchestration of different modules to achieve
 // a cohesive execution flow.
-package review
+package logic
